# Optimize date fallback methods