Hey there,

I was attempting to use your code on a LOT (1-10 million rows) and found that it was not performant enough. Upon digging in, I discoverd the culprit is the date fallback code, which is doing a linear scan across ~5000 dates each time the caller asks for a date which does not have data.

Since any holes in the data tend to be only a few days at most, I made a simple optimization to first try looking 5 days in either direction before falling back to the full linear scan. I'm now able to run on 10 million rows in a matter of minutes.

Thanks,
Greg