@alexprengere 2016-04-22
Hello and thanks!
I did a naive implementation which indeed was not smart at all regarding fallbacks. Your implementation looks good, and thanks for the tests as well :wink:. I am going to merge this.

I thought a bit more about it, and I think we can even go a step further. We could compute the holes at load time once, then any call to the library would be in constant time. Since the fallback options are passed during initialization, we wouldn't have to break the API for this:

* if `fallback_on_missing_rate` is true, all dates between the first and last available date are filled with the closest rate during initialization
* if a date is requested before the first available or after the last available, and `fallback_on_wrong_date` is true, we use the first or last date (depending) at request time

Does this sound good to you? I can probably implement this quickly.