@gpascale 2016-04-22
Thanks Alex,

Yep, that makes a lot of sense. In fact, I thought about doing it that way,
but I needed to get something done quickly so I could unblock my other
project. Look forward to seeing your implementation.

Greg



On Fri, Apr 22, 2016 at 7:31 AM Alex Prengère <notifications@github.com>
wrote:

> Merged #2 <https://github.com/alexprengere/currencyconverter/pull/2>.
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly or view it on GitHub
> <https://github.com/alexprengere/currencyconverter/pull/2#event-637774998>
>