@alexprengere 2016-04-28
Sooo it took way longer than I thought, but I think it is OK now (pushed on master). The problem was that the naive implementation took forever (for each missing rate, look up the closest one: quadratic), so I had to do it with a bit more code but in linear time.

Now if `fallback_on_missing_rate` is `True`, between the first and last date available for the currency, all rates will be linearly interpolated from the closest available rates. Outside the first/last date available for the currency, `ValueError` is raised unless `fallback_on_wrong_date` is `True`, in this case we either the first or last date as fallback. With those options turned on, the load time is now about 3 seconds.

I made many other small changes to the code base (mainly refactoring and optimizations), so I hope I did not break anything, and I hope your code will run much faster now :wink:. I updated the doc as well, I am waiting for your feedback before uploading a new package on PyPI. Thanks!