@gpascale 2016-04-28
This is awesome. I'll try it out today!
On Thu, Apr 28, 2016 at 7:04 AM Alex Prengère <notifications@github.com>
wrote:

> I ran some benchmark, testing a convert() call:
>
>    - before your contribution, it was 8µs for a known rate and 15ms for
>    an unknown rate (full scan)
>    - after your contribution, it was 8µs for a known rate and 35µs for an
>    unknown rate
>    - now it typically takes 6µs (whether or not it was a unknown rate)
>
> So on a data set of 10 000 000 rows I would expect, depending on the
> number of unknown rate you get, between and 0 and 4 minutes of speedup [image:
> :smile:].
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly or view it on GitHub
> <https://github.com/alexprengere/currencyconverter/pull/2#issuecomment-215433529>
>